메인페이지 및 데일리케어 관련 기능들 나누기
메인페이지 및 데일리케어 관련 기능들 나누기
